- 자료형 변경 - 메서드 -> util - 정산조회 추가 - 전기차 정산 대비 - 기능 구현 완료

+++ src/main/java/controller/ParkingController.java
@@ -9,7 +9,6 @@ public class ParkingController {
 
     public final ParkingService parkingService = new ParkingService();
 
-    // TODO: 입차
     public void innerCar(String type, int capacity, String carNumber) {
         try{
             parkingService.innerCar(type, capacity, carNumber);
@@ -18,18 +17,15 @@ public class ParkingController {
         }
     }
 
-    // TODO: 출차, 정산
     public void exitCar(String carNumber) throws Exception {
         parkingService.exitCar(carNumber);
     }
 
-    // TODO: 차량 조회
     public Map<String, CarInfo> getCarList(){
         return parkingService.getCarList();
     }
 
-    // TODO: 수입 조회
-    public Long getTotalSettlement(){
+    public double getTotalSettlement(){
         return parkingService.getTotalSettlement();
     }
 }
